Added support for YunoHost
Hi 
I have packaged a Firefly  III app for YunoHost. It has install,remove,update,backup,restore script along with multi-instance(can be installed multiple times on different domains on a single server). The app is tested and works well. 
The CL passing will take a day to show up on the app page.
Thanks for coding this app.